Fixed Track Removal
Had an issue where attempting to remove a track from the playlist removed every track besides the one that you were attempting to remove. This is fixed by just calling the .splice method, instead of passing what is returned to the state of it, since it returns what is removed from the array.